backend/requirements.txt
asgiref==3.7.2
Babel==2.12.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.20
django-cors-headers==4.2.0
django-environ==0.10.0
django-extra-views==0.14.0
django-haystack==3.2.1
django-oscar==3.2
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-treebeard==4.7
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
factory-boy==3.2.1
Faker==19.3.1
idna==3.4
inflection==0.5.1
phonenumbers==8.13.19
Pillow==10.0.0
purl==1.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.4
stripe==6.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
